+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+# **
+
+* Improved performance :
+  - Better cache of LambdaWithParams
+  - Skip check of inversability when inverting matrices with LU method
+
+* Integrate PR #85 : Fix and improve interpolate with units 
+
 # 1.3.2 
 
 * Fix bug #82 impact computation by axis not working. For nested axes.